+++ lib/cache_x.dart
@@ -1,6 +1,11 @@
 library cache_x;
 
+import 'dart:typed_data';
+
 import 'package:cache_x/src/cacheXWorker.dart';
+import 'package:flutter/material.dart';
+import 'package:helpers/helpers.dart';
 
 part 'src/cacheXCore.dart';
 part 'src/exception.dart';
+part 'src/debugLog.dart';