+++ test/cache_x_test.dart
@@ -1,1 +1,5 @@
-void main() {}
+import 'package:flutter_test/flutter_test.dart';
+
+void main() {
+  test('Test if the file is saving and retriving properly', () async {});
+}